PrepareTags moved to .Core
