+++ Sources/Ditch.Golos.Tests/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Cryptography.ECDSA.Secp256k1" version="1.0.1" targetFramework="net46" />
-  <package id="Ditch.Core" version="3.0.0" targetFramework="net46" />
+  <package id="Ditch.Core" version="3.0.1" targetFramework="net46" />
   <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="net46" />
   <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="net46" />
   <package id="NETStandard.Library" version="1.6.1" targetFramework="net46" />
