+++ Sources/Ditch.Steem/Operations/Post/CommentOperation.cs
@@ -1,5 +1,4 @@
-﻿using System.Text.RegularExpressions;
-using Ditch.Steem.Helpers;
+﻿using Ditch.Steem.Helpers;
 using Newtonsoft.Json;
 
 namespace Ditch.Steem.Operations.Post
@@ -10,15 +9,6 @@ namespace Ditch.Steem.Operations.Post
     [JsonObject(MemberSerialization.OptIn)]
     public class CommentOperation : BaseOperation
     {
-        /// <summary>
-        /// @"[_\s\.]+"
-        /// </summary>
-        protected static readonly Regex WordDelimiters = new Regex(@"[_\s\.]+");
-        /// <summary>
-        /// @"[^a-z0-9-]+"
-        /// </summary>
-        protected static readonly Regex PermlinkNotSupportedCharacters = new Regex(@"[^a-z0-9-]+", RegexOptions.IgnoreCase);
-
         public override string TypeName => "comment";
 
         public override OperationType Type => OperationType.Comment;
