+++ Sources/Ditch.Steem/Operations/Post/PostOperation.cs
@@ -26,30 +26,10 @@ namespace Ditch.Steem.Operations.Post
         {
             text = text.Trim();
             text = text.ToLower();
-            text = WordDelimiters.Replace(text, "-");
+            text = Transliteration.WordDelimiters.Replace(text, "-");
             text = Transliteration.ToEng(text);
-            text = PermlinkNotSupportedCharacters.Replace(text, string.Empty);
+            text = Transliteration.PermlinkNotSupportedCharacters.Replace(text, string.Empty);
             return $"{text}-{DateTime.UtcNow:yyyy-MM-dd-HH-mm-ss}";
-        }
-
-
-        /// <summary>
-        /// Generate a valid tags
-        /// </summary>
-        /// <param name="tags"></param>
-        public static void PrepareTags(string[] tags)
-        {
-            for (var index = 0; index < tags.Length; index++)
-            {
-                var tag = tags[index];
-                tag = tag.Trim();
-                tag = tag.ToLower();
-                var translit = Transliteration.ToEng(tag);
-                tag = translit.Equals(tag) ? translit : $"ru--{translit}";
-                tag = WordDelimiters.Replace(tag, "-");
-                tag = PermlinkNotSupportedCharacters.Replace(tag, string.Empty);
-                tags[index] = tag;
-            }
         }
     }
 }
